🛠️ Refactor: Extract magic numbers and isolate error handling package
Replaced magic numbers `64*1024` and `1 * time.Second` with named constants in `main.go`.
Relocated `errors.go` to `internal/apperror` package to enforce domain grouping.
