+++ internal/apperror/errors.go
@@ -1,4 +1,4 @@
-package main
+package apperror
 
 import (
 	"log"
