+++ main.go
@@ -14,12 +14,19 @@ import (
 	"time"
 
 	"github.com/davecgh/go-spew/spew"
+	"github.com/lucasew/ncgi/internal/apperror"
 	"github.com/phayes/freeport"
 )
 
 func init() {
 	_ = spew.Config
 }
+
+const (
+	defaultBufSize        = 64 * 1024
+	subprocessIdleDelay   = 1 * time.Second
+	subprocessWarmupDelay = 1 * time.Second
+)
 
 var bufsize int
 var script string
@@ -30,7 +37,7 @@ func main() {
 	defer cancel()
 	flag.StringVar(&script, "s", "./example", "script to be CGIed")
 	flag.IntVar(&port, "p", 0, "port to listen")
-	flag.IntVar(&bufsize, "b", 64*1024, "buffer size")
+	flag.IntVar(&bufsize, "b", defaultBufSize, "buffer size")
 	flag.Parse()
 	if port == 0 {
 		port = freeport.GetPort()
@@ -40,19 +47,19 @@ func main() {
 	defer func() {
 		err := server.Shutdown(ctx)
 		if err != nil {
-			ReportError(err, nil)
+			apperror.ReportError(err, nil)
 		}
 	}()
 	go func() {
 		err := server.ListenAndServe()
 		if err != nil && err != http.ErrServerClosed {
-			ReportFatal(err, nil)
+			apperror.ReportFatal(err, nil)
 		}
 	}()
 
 	err := handleSubprocess(ctx, flag.Args()...)
 	if err != nil {
-		ReportFatal(err, nil)
+		apperror.ReportFatal(err, nil)
 	}
 }
 
@@ -60,7 +67,7 @@ func handleSubprocess(ctx context.Context, args ...string) error {
 	var err error
 	if len(args) == 0 {
 		for {
-			time.Sleep(1 * time.Second)
+			time.Sleep(subprocessIdleDelay)
 		}
 	}
 	args[0], err = exec.LookPath(args[0])
@@ -74,7 +81,7 @@ func handleSubprocess(ctx context.Context, args ...string) error {
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
 	cmd.Stdin = os.Stdin
-	time.Sleep(1 * time.Second)
+	time.Sleep(subprocessWarmupDelay)
 	return cmd.Run()
 }
 
@@ -85,11 +92,11 @@ type CGIHandler struct {
 func NewCGIHandler(script string) http.Handler {
 	p, err := exec.LookPath(script)
 	if err != nil {
-		ReportFatal(err, map[string]interface{}{"script": script})
+		apperror.ReportFatal(err, map[string]interface{}{"script": script})
 	}
 	p, err = filepath.Abs(p)
 	if err != nil {
-		ReportFatal(err, map[string]interface{}{"path": p})
+		apperror.ReportFatal(err, map[string]interface{}{"path": p})
 	}
 	log.Printf("Initializing CGI handler on folder '%s'...", p)
 	return CGIHandler{p}
@@ -131,7 +138,7 @@ func (c CGIHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	cmd.Stdin = r.Body
 	stdout, err := cmd.StdoutPipe()
 	if err != nil {
-		ReportError(err, nil)
+		apperror.ReportError(err, nil)
 		fmt.Fprint(w, err.Error())
 		return
 	}
@@ -140,7 +147,7 @@ func (c CGIHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		if cmd.Process != nil {
 			err := cmd.Process.Kill()
 			if err != nil {
-				ReportError(err, nil)
+				apperror.ReportError(err, nil)
 			}
 		}
 	}()
@@ -148,7 +155,7 @@ func (c CGIHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	cmd.Stderr = os.Stderr
 	err = cmd.Start()
 	if err != nil {
-		ReportError(err, map[string]interface{}{"script": c.script})
+		apperror.ReportError(err, map[string]interface{}{"script": c.script})
 
 		// w.WriteHeader(500)
 		fmt.Fprint(w, err.Error())
@@ -165,13 +172,13 @@ func (c CGIHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 				return
 			}
 			if err != nil {
-				ReportError(err, nil)
+				apperror.ReportError(err, nil)
 				fmt.Fprint(w, err.Error())
 				return
 			}
 			_, err = w.Write(buf[:sz])
 			if err != nil {
-				ReportError(err, nil)
+				apperror.ReportError(err, nil)
 				fmt.Fprint(w, err.Error())
 				return
 			}
